refactor: remove redundant 'github' prefix from filenames
- Rename github-checkout.ts to checkout.ts
- Rename github-api.ts to api.ts
- Rename pr-checkout.ts to pr.ts
- Rename issue-checkout.ts to issue.ts
- Update all import paths and test file names accordingly

Since these files are already in the packages/github directory,
the 'github' prefix in filenames is redundant.
