+++ packages/github/src/checkout/issue.test.js
@@ -1,6 +1,6 @@
 import { equal } from "node:assert/strict";
 import { describe, it } from "node:test";
-import { checkoutIssue } from "./issue-checkout.ts";
+import { checkoutIssue } from "./issue.ts";
 
 describe("checkoutIssue", () => {
   it("should export checkoutIssue function", () => {
