+++ packages/github/src/checkout/issue.ts
@@ -4,23 +4,31 @@ import {
 } from "@aku11i/phantom-core";
 import { getGitRoot } from "@aku11i/phantom-git";
 import { isErr } from "@aku11i/phantom-shared";
-import type { GitHubPullRequest } from "../github-api.ts";
+import type { GitHubIssue } from "../api.ts";
 
-export async function checkoutPullRequest(
-  pullRequest: GitHubPullRequest,
+export async function checkoutIssue(
+  issue: GitHubIssue,
   number: string,
+  base?: string,
 ): Promise<void> {
+  if (issue.pull_request) {
+    throw new Error(
+      `#${number} is already linked to a pull request. Use the PR number instead.`,
+    );
+  }
+
   const gitRoot = await getGitRoot();
-  const worktreeName = `pr-${number}`;
+  const worktreeName = `issue-${number}`;
+  const branchName = `issue-${number}`;
 
   const result = await createWorktreeCore(gitRoot, worktreeName, {
-    branch: pullRequest.head.ref,
-    base: `origin/${pullRequest.head.ref}`,
+    branch: branchName,
+    base,
   });
 
   if (isErr(result)) {
     if (result.error instanceof WorktreeAlreadyExistsError) {
-      console.log(`Worktree for PR #${number} is already checked out`);
+      console.log(`Worktree for issue #${number} is already checked out`);
       return;
     }
     throw result.error;
