+++ packages/github/src/checkout/pr.ts
@@ -4,31 +4,23 @@ import {
 } from "@aku11i/phantom-core";
 import { getGitRoot } from "@aku11i/phantom-git";
 import { isErr } from "@aku11i/phantom-shared";
-import type { GitHubIssue } from "../github-api.ts";
+import type { GitHubPullRequest } from "../api.ts";
 
-export async function checkoutIssue(
-  issue: GitHubIssue,
+export async function checkoutPullRequest(
+  pullRequest: GitHubPullRequest,
   number: string,
-  base?: string,
 ): Promise<void> {
-  if (issue.pull_request) {
-    throw new Error(
-      `#${number} is already linked to a pull request. Use the PR number instead.`,
-    );
-  }
-
   const gitRoot = await getGitRoot();
-  const worktreeName = `issue-${number}`;
-  const branchName = `issue-${number}`;
+  const worktreeName = `pr-${number}`;
 
   const result = await createWorktreeCore(gitRoot, worktreeName, {
-    branch: branchName,
-    base,
+    branch: pullRequest.head.ref,
+    base: `origin/${pullRequest.head.ref}`,
   });
 
   if (isErr(result)) {
     if (result.error instanceof WorktreeAlreadyExistsError) {
-      console.log(`Worktree for issue #${number} is already checked out`);
+      console.log(`Worktree for PR #${number} is already checked out`);
       return;
     }
     throw result.error;
