+++ packages/github/src/index.ts
@@ -1,2 +1,2 @@
-export * from "./github-checkout.ts";
-export type { GitHubPullRequest, GitHubIssue } from "./github-api.ts";
+export * from "./checkout.ts";
+export type { GitHubPullRequest, GitHubIssue } from "./api.ts";
